compilation-database: add a portable sed invocation
The given `sed` invocation to merge compilation database entries
generated by clang's `-MJ` flag is not POSIX compliant, as such
it does not work on macOS or other BSDs.

Add an alternative using ANSI-C Quoting to quote the newline, which
should work in Bash, Zsh and ksh.

Reference: https://stackoverflow.com/a/24299845